Added a bunch of small changes to Graph and Data
Made it so that the data will automatically have the correct ID number
based on the database length and added the ability to limit how far back
in the database graph data is pulled from. Currently at 20 entries for
testing. Added a new abstraction which returns the number of rows in a
table and edited two other abstractions to allow for extra parameters
needed. It's a little messy now as there are IF/Else statements, but
those can be removed once a standard is set.